feat: hot loads modules
a big spike that prob adds some breaking changes, but the main feature this
adds is the ability to create the entire ecto and absinthe app on the
fly, and re-load it when the database changes. it also includes some
macros for extending/injecting code into the auto generated code.

All the generated code is written to `.exs` files that aren't compiled
but help you see what's being generated.